requirements.txt
pandas==2.2.0
qrcode[pil]==7.4.2
pillow==10.2.0
python-dotenv==1.0.1
